Implements sign out and refactors app.rb into controllers
