+++ app/controllers/peeps.rb
@@ -1,0 +1,17 @@
+class Chitter < Sinatra::Base
+
+  get '/peeps' do
+    @peeps = Peep.all
+    erb :'peeps/index'
+  end
+
+  get '/peeps/new' do
+    erb :'peeps/new'
+  end
+
+  post '/peeps' do
+    peep = Peep.create(message: params[:message], time_posted: Time.now)
+  redirect to '/peeps'
+  end
+
+end
