+++ app/controllers/sessions.rb
@@ -1,0 +1,23 @@
+class Chitter < Sinatra::Base
+
+  get '/sessions/new' do
+    erb :'sessions/new'
+  end
+
+  post '/sessions' do
+    user = User.authenticate(params[:user_name], params[:password])
+    if user
+      session[:user_id] = user.id
+      redirect to '/peeps'
+    else
+      flash.now[:notice] = 'Login Failed. Please re-enter your details.'
+      erb :'sessions/new'
+    end
+  end
+
+  delete '/sessions' do
+    session[:user_id] = nil
+    redirect to '/'
+  end
+
+end
