+++ app/views/index.erb
@@ -1,7 +1,1 @@
-<h1>Welcome to Chitter!</h1>
-<form action='/users/new' method='get'>
-  <input type='submit' value='Sign Up'>
-</form>
-<form action='/sessions/new' method='get'>
-  <input type='submit' value='Sign In'>
-</form>
+
