+++ app/views/peeps/index.erb
@@ -1,5 +1,3 @@
-<h1>Chitter</h1>
-Hello <%= current_user.name %>
 <ul>
   <% @peeps.reverse.each do |peep| %>
     <li>
@@ -13,9 +11,3 @@ Hello <%= current_user.name %>
     </li>
   <% end %>
 </ul>
-
-<div class="Add_a_new_Peep">
-  <form action='/peeps/new' method='get'>
-    <input type='submit' value='Add a new Peep'>
-  </form>
-</div>
