+++ app/views/peeps/new.erb
@@ -1,4 +1,4 @@
-<h1>Hello <%= current_user.name %>. Please enter a Peep</h1>
+<h1>Please enter a Peep</h1>
 <form action='/peeps' method='post'>
   <label for='peep'>
     Peep
