+++ app/views/sessions/new.erb
@@ -1,5 +1,3 @@
-<h1>Welcome back to Chitter!</h1>
-
 <form action='/sessions' method='post'>
   <label for='user_name'>
     User Name
