+++ spec/features/user_sign_in_spec.rb
@@ -1,11 +1,12 @@
 feature 'Sign In' do
   scenario 'I can sign in as a User' do
+    User.create(name: 'Tom Spencer', user_name:'tomspencer', password: 'tomspencer')
     sign_in
-    expect(page).to have_content('Hello Tom Spencer')
+    expect(page).to have_content('Welcome Tom Spencer')
   end
 
   scenario 'I see a fail message if I sign in incorrectly' do
     sign_in_with_wrong_password
-  expect(page).to have_content('Welcome back to Chitter!')
+  expect(page).to have_content('Login Failed')
   end
 end
